refactor: move document imples
