+++ src/doc-entity/index.ts
@@ -1,0 +1,2 @@
+export { DefaultDocumentEntity } from "./default-document";
+export { BabylonDocumentEntity } from "./babylon-document";
